_archive/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2==2.9.9
chromadb>=0.5.0
google-generativeai==0.3.2
python-multipart==0.0.6
pandas>=2.0.0
python-dotenv==1.0.0
pydantic>=2.5.0
langchain==0.0.350
langchain-google-genai==0.0.5
python-docx==1.1.0
PyPDF2==3.0.1

# Monitoring and Observability
prometheus-client==0.19.0
psutil==5.9.6
redis==5.0.1
asyncpg==0.29.0
aiohttp==3.9.1
sentry-sdk==1.38.0
python-json-logger==2.0.7
structlog==23.2.0

# Additional dependencies for monitoring
requests==2.31.0
jinja2==3.1.2
